requirements.txt
ffmpeg-python==0.2.0
future==0.18.3
numpy==1.26.2
opencv-python==4.8.1.78
OpenTimelineIO==0.15.0
pyaaf2==1.6.0
PySide6==6.3.1
PySide6-Addons==6.3.1
PySide6-Essentials==6.3.1
shiboken6==6.3.1
